Create stat card component
