+++ graylog2-web-interface/src/components/collectors/common/index.ts
@@ -14,5 +14,4 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-export { default as StatCard } from './StatCard';
 export { default as CollectorsPageNavigation } from './CollectorsPageNavigation';
