+++ graylog2-web-interface/src/components/collectors/overview/CollectorsOverview.tsx
@@ -17,6 +17,7 @@
 import * as React from 'react';
 import { useState } from 'react';
 import styled, { css } from 'styled-components';
+import type { ColorVariant } from '@graylog/sawmill';
 
 import { Alert, Input } from 'components/bootstrap';
 import { Spinner } from 'components/common';
@@ -25,6 +26,7 @@ import Routes from 'routing/Routes';
 import { TELEMETRY_EVENT_TYPE } from 'logic/telemetry/Constants';
 import type { CollectorStats } from 'components/collectors/types';
 import useFeature from 'hooks/useFeature';
+import StatCard from 'components/common/StatCard/StatCard';
 
 import FleetCardsGrid from './FleetCardsGrid';
 import RecentActivity from './RecentActivity';
@@ -32,7 +34,6 @@ import FirstOnboarding from './FirstOnboarding';
 
 import { useCollectorStats, useFleetsBulkStats } from '../hooks';
 import useSendCollectorsTelemetry from '../hooks/useSendCollectorsTelemetry';
-import StatCard, { type Variant as StatCardVariant } from '../common/StatCard';
 
 const StatsRow = styled.div(
   ({ theme }) => css`
@@ -74,7 +75,7 @@ const StatsSection = ({ stats }: { stats: CollectorStats }) => {
   const emitStatCard = (
     card: 'instances' | 'online' | 'offline' | 'fleets',
     value: number,
-    variant: StatCardVariant,
+    variant: ColorVariant,
     navigates_to: 'instances' | 'instances-online' | 'instances-offline' | 'fleets',
   ) => {
     sendTelemetry(TELEMETRY_EVENT_TYPE.COLLECTORS.OVERVIEW.STAT_CARD_CLICKED, {
