+++ graylog2-web-interface/src/components/common/StatCard/StatCard.test.tsx
