+++ graylog2-web-interface/src/components/common/index.tsx
@@ -162,3 +162,4 @@ export { default as Toggle } from './Switch';
 export { default as SearchFiltersFormControls } from './SearchFiltersFormControls';
 export { default as ModalButtonToolbar } from './ModalButtonToolbar';
 export { default as AccessibleCard } from './AccessibleCard';
+export { default as StatCard } from  './StatCard/StatCard';
